requirements.txt
Jinja2>=3.1
openpyxl>=3.0
pandas>=1.4
python-dotenv>=0.20
configargparse>=1.5